tasks/classified/requirements.txt
qrcode==7.3.1
Pillow
